Started working on the menubar after doing some research on how best to do it. It appears that Material-ui is really nice for this sort of thing. I had used it previously and forgot. It's in addition to and sits on top of react to add additional functionality and makes things look cleaner.
Now what I'll do is get it to show the menu icon on the right-hand side
as well as insert all of the links that we'll need for navigating from
the home page, the about me section, and chat.

However, now that we found a component online that may be able to
handle chat for us, we may go with that instead which would be a
floating chat box, not necessitating its own window.