Added a router, having really touched it yet. I did however flesh out the menu a bit and got the events to work correctly after wrestling with them for a while.
The next step is to get the routes to work with the MenuItem's. I think
that I may be having difficulty because I'm not looking to place the
navigation in the right place. Maybe I do not even need to use MenuItems
and instead need Routes. I dunno, I'm tired.